requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-multipart
python-dotenv
